meeting-transcriber/backend/pyproject.toml
[project]
name = "meeting-transcriber"
version = "0.1.0"
description = "Real-time meeting transcription with Slack summaries"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.7.0",
    "alembic>=1.13",
    "asyncpg>=0.29",
    "fastapi>=0.115.0",
    "openai>=2.31.0",
    "psycopg[binary]>=3.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "slack-sdk>=3.41.0",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.34.0",
    "websockets>=14.0",
    "bcrypt>=4.0.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
]
